+++ theme/src/typescript/admin/blocks/index.ts
@@ -1,0 +1,6 @@
+import speaker from './speaker';
+
+(() => {
+  // Enabling on load.
+  speaker.register();
+})();