+++ theme/template/functions.php
@@ -1,3 +1,2 @@
 <?php
-include_once dirname(__FILE__) . '/functions/categories.php';
-include_once dirname(__FILE__) . '/functions/features.php';
+include_once dirname(__FILE__) . '/functions/index.php';