+++ theme/template/pages/speakers.php
@@ -5,13 +5,17 @@
 
 $cats = [
   SPEAKER_SPONSORED => get_category_by_slug(SPEAKER_SPONSORED),
-  SPEAKER_COMMUNITY => get_category_by_slug(SPEAKER_COMMUNITY),
+  SPEAKER_COMMUNITY => get_category_by_slug(SPEAKER_COMMUNITY)
 ];
 
 // Retrieve all posts.
 $posts = [
-  SPEAKER_SPONSORED => get_posts(['category' => $cats[SPEAKER_SPONSORED]->cat_ID]),
-  SPEAKER_COMMUNITY => get_posts(['category' => $cats[SPEAKER_COMMUNITY]->cat_ID]),
+  SPEAKER_SPONSORED => get_posts([
+    'category' => $cats[SPEAKER_SPONSORED]->cat_ID
+  ]),
+  SPEAKER_COMMUNITY => get_posts([
+    'category' => $cats[SPEAKER_COMMUNITY]->cat_ID
+  ])
 ];
 ?>
 <section class="wrapper">
@@ -25,17 +29,17 @@ $posts = [
     </div>
 
     <div class="row speaker__list">
-      <?php foreach($posts[SPEAKER_SPONSORED] as $p): ?>
+      <?php foreach ($posts[SPEAKER_SPONSORED] as $p): ?>
         <div class="col-6 col-sm-3">
-          <a class="speaker__summary" href="<?= get_permalink($p->ID); ?>">
+          <a class="speaker__summary" href="<?= get_permalink($p->ID) ?>">
             <img class="speaker__summary-image img-fluid" src="<?= get_template_directory_uri() ?>/assets/images/placeholder.png">
 
             <p class="speaker__summary-name">
-              <?= $p->post_title; ?>
+              <?= $p->post_title ?>
             </p>
 
             <p class="speaker__summary-message">
-              <?= wp_strip_all_tags($p->post_content); ?>
+              <?= wp_strip_all_tags($p->post_content) ?>
             </p>
           </a>
         </div>
@@ -53,17 +57,17 @@ $posts = [
     </div>
 
     <div class="row speaker__list">
-      <?php foreach($posts[SPEAKER_COMMUNITY] as $p): ?>
+      <?php foreach ($posts[SPEAKER_COMMUNITY] as $p): ?>
         <div class="col-6 col-sm-3">
-          <a class="speaker__summary" href="<?= get_permalink($p->ID); ?>">
+          <a class="speaker__summary" href="<?= get_permalink($p->ID) ?>">
             <img class="speaker__summary-image img-fluid" src="<?= get_template_directory_uri() ?>/assets/images/placeholder.png">
 
             <p class="speaker__summary-name">
-              <?= $p->post_title; ?>
+              <?= $p->post_title ?>
             </p>
 
             <p class="speaker__summary-message">
-              <?= wp_strip_all_tags($p->post_content); ?>
+              <?= wp_strip_all_tags($p->post_content) ?>
             </p>
           </a>
         </div>